requirements.txt
autopep8==1.5.4
click==7.1.2
dlib==19.21.1
face-recognition==1.3.0
face-recognition-models==0.3.0
numpy==1.19.5
Pillow==8.1.0
pycodestyle==2.6.0
toml==0.10.2
